+++ __init__.py
@@ -1,3 +1,5 @@
-from .wmcpy import Datapack, create_file, create_pack_meta, Raycast, import_from_file
+from .datapack import Datapack
+from .files import *
+from .utils import Raycast
 VERSION = "0.0.1"
 AUTHOR = "Vianpyro"
